refactor: convert cypress config to ESM
